% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-stats.R
\name{aov}
\alias{aov}
\title{data.frame-first formula-second method for \code{\link[stats]{aov}}}
\usage{
aov(data, formula, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[stats]{aov}}.
}
\examples{
fit <- aov(mtcars, hp ~ am * cyl)
summary(fit)
predict(fit, mtcars)
}
\seealso{
\code{\link[stats]{aov}}
}
