% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_alt_args}
\alias{check_alt_args}
\title{Check if argument(s) are given as alternatives to another}
\usage{
check_alt_args(..., target, alts, stop = TRUE)
}
\arguments{
\item{...}{Named arguments. Typically passed from another function.}

\item{target}{String of variable for which alternatives may be present.}

\item{alts}{Character vector of arguments that may act as alternatives to the target}

\item{stop}{Boolean. If True, \code{\link[base]{stop}}() is called with an
error message. If False, \code{\link[base]{warning}}() is called with an
error message.}
}
\description{
Check for certain arguments that, if present, will result in function
producing an error and stopping the function if desired.
}
\examples{
check_alt_args(a = 10, target = "a", alts = c("b"))
check_alt_args(b = 10, target = "a", alts = c("b"))
check_alt_args(b = 10, target = "a", alts = c("b"), stop = FALSE)
}
