% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_alt_data}
\alias{check_alt_data}
\title{Wrapper for \code{\link{check_alt_arg}} with data defaults}
\usage{
check_alt_data(..., target = "data", alts = c("newx", "newdata"),
  stop = TRUE)
}
\description{
Wrapper for \code{\link{check_alt_arg}} with data defaults
}
\examples{
check_alt_data(data = 10)
\dontrun{
check_alt_data(newx = 10)
check_alt_data(newdata = 10)
}
}
