% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_pkg}
\alias{check_pkg}
\title{Check for required package}
\usage{
check_pkg(pkg_name)
}
\arguments{
\item{pkg_name}{Character of the package name to check}
}
\description{
Check if package is locally installed and call `stop()` with an error message
if not
}
\examples{
check_pkg("stats")
check_pkg("dplyr")
check_pkg("zzz")
}
