% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{coerce_args}
\alias{coerce_args}
\title{Coerce arguments to right object class}
\usage{
coerce_args(data, formula)
}
\arguments{
\item{data}{Object that can be coerced to a data frame}

\item{formula}{Object that can be coerced to a formula}
}
\description{
\code{coerce_args} is used to ensure that the data and formula arguments
provided to twidlr model functions are coerced to a data frame and formula
before further processing. This makes it possible for remaining function code
to assume these object classes.
}

