% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-quantreg.R
\name{crq}
\alias{crq}
\alias{predict.crq}
\title{data.frame-first formula-second method for \code{\link[quantreg]{crq}}}
\usage{
crq(data, formula, ...)

\method{predict}{crq}(object, data, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{\code{\link[stats]{formula}} describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[quantreg]{crq}}.
}
\examples{
set.seed(1968)
n <- 200
x <- rnorm(n)
y <- 5 + x + rnorm(n)
c <- 4 + x + rnorm(n)
d <- (y > c)
dat <- data.frame(y = y,x = x,c = c,d = d)
rm(x,y,c,d)
f2 <- crq(dat,survival::Surv(pmax(y,c), d, type = "left") ~ x,method = "Portnoy")

# Help page for function being twiddled
?quantreg::crq
}
\seealso{
\code{\link[quantreg]{crq}}
}
