% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-glmnet.R
\name{cv.glmnet}
\alias{cv.glmnet}
\alias{predict.cv.glmnet}
\title{data.frame-first formula-second method for \code{\link[glmnet]{cv.glmnet}}}
\usage{
cv.glmnet(data, formula, ...)

\method{predict}{cv.glmnet}(object, data, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[glmnet]{cv.glmnet}}().
}
\examples{
fit <- cv.glmnet(mtcars, hp ~ .)
predict(fit, mtcars, s = "lambda.min")
}
\seealso{
\code{\link[glmnet]{cv.glmnet}}
}

