% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{eval_fun}
\alias{eval_fun}
\title{Evaluate a function and list of arguments}
\usage{
eval_fun(.f, args)
}
\arguments{
\item{.f}{A function}

\item{args}{Named list of arguments}
}
\description{
Evaluate a function and a named list of corresponding arguments
}
