% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-gamlss.R
\name{gamlss}
\alias{gamlss}
\alias{predict.gamlss}
\title{data.frame-first formula-second method for \code{\link[gamlss]{gamlss}}}
\usage{
gamlss(data, formula, ...)

\method{predict}{gamlss}(object, data, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[gamlss]{gamlss}}.
}
\details{
This function attaches the training data to the generated model, because the
default gamlss.predict() function requires access to the original data in
order to make predictions. Please be aware that memory issues can arise when
working with large data sets and or when generating a large number of models.
}
\examples{
fit <- gamlss(mtcars, vs ~ hp + wt, family = "BI")
summary(fit)
predict(fit, mtcars)
}
\seealso{
\code{\link[gamlss]{gamlss}}
}
