% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm}
\alias{glm}
\title{data.frame-first formula-second method for \code{\link[stats]{glm}}}
\usage{
glm(data, formula, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{an object of class \code{"\link{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{...}{additional arguments to pass to \code{\link[stats]{glm}}}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[stats]{glm}}.
}
\examples{
fit <- glm(mtcars, vs ~ hp + wt, family = binomial())
summary(fit)

# Help page for function being twiddled
?stats::glm
}
\seealso{
\code{\link[stats]{glm}}
}
