% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet}
\alias{glmnet}
\title{data.frame-first formula-second method for \code{\link[glmnet]{glmnet}}}
\usage{
glmnet(data, formula, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model.}

\item{formula}{A formula of the form \code{y ~ x1 + x2 + ... + xi}}

\item{...}{additional arguments to pass to \code{\link[glmnet]{glmnet}}()}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[glmnet]{glmnet}}().
}
\examples{
glmnet(mtcars, hp ~ .)

# Help page for function being twiddled
?glmnet::glmnet
}
\seealso{
\code{\link[glmnet]{glmnet}}
}
