% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm}
\alias{lm}
\title{data.frame-first formula-second method of stats::lm}
\usage{
lm(data, formula, ...)
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[stats]{lm}}().
}
