% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-lme4.R
\name{lmer}
\alias{lmer}
\alias{predict.merMod}
\title{data.frame-first formula-second method for \code{\link[lme4]{lmer}}}
\usage{
lmer(data, formula, ...)

\method{predict}{merMod}(object, data, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[lme4]{lmer}}
}
\examples{
fit <- lmer(airquality, Ozone ~ Wind + (Wind | Month))
summary(fit)

# Help page for function being twiddled
?lme4::lmer
}
\seealso{
\code{\link[lme4]{lmer}}
}

