% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{model_as_xy}
\alias{model_as_xy}
\title{Convert data frame and model \code{\link[stats]{formula}} to input matrix and
output vector list}
\usage{
model_as_xy(data, formula)
}
\arguments{
\item{data}{A data frame}

\item{formula}{A \code{\link[stats]{formula}}}
}
\value{
Two-element list containing input variables as a matrix \code{x}, and
  output variable as a vector \code{y}
}
\description{
Convert a data frame and model \code{\link[stats]{formula}} to a list
containing the input variables as a matrix \code{x} and the output variable
as a vector \code{y}. These structures are commonly used by various modelling
functions such as \code{\link[glmnet]{glmnet}}.
}
\details{
All operators made available by \code{\link[stats]{formula}} are allowed.
}
\examples{
model_as_xy(mtcars, hp ~ mpg)
model_as_xy(mtcars, hp ~ .)
model_as_xy(mtcars, hp ~ mpg * hp)
model_as_xy(mtcars, ~ .)
}
\seealso{
\code{\link[stats]{formula}}
}
