% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-e1071.R
\name{naiveBayes}
\alias{naiveBayes}
\alias{predict.naiveBayes}
\title{data.frame-first formula-second method for \code{\link[e1071]{naiveBayes}}}
\usage{
naiveBayes(data, formula, ...)

\method{predict}{naiveBayes}(object, data, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[e1071]{naiveBayes}}.
}
\examples{

data <- mtcars
data$cyl_gr_6 <- ifelse(data$cyl > 6, 1, 0)

fit <- naiveBayes(data, cyl_gr_6 ~ .)
summary(fit)

# Help page for function being twiddled
?e1071::naiveBayes

}
\seealso{
\code{\link[e1071]{naiveBayes}}
}

