% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlrq.R
\name{nlrq}
\alias{nlrq}
\alias{predict.nlrq}
\title{data.frame-first formula-second method for \code{\link[quantreg]{nlrq}}}
\usage{
nlrq(data, formula, ...)

\method{predict}{nlrq}(object, data, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{\code{\link[stats]{formula}} describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[quantreg]{nlrq}}.
}
\examples{
Dat <- NULL
Dat$x <- rep(1:25, 20)
set.seed(1)
Dat$y <- SSlogis(Dat$x, 10, 12, 2) * rnorm(500, 1, 0.1)

fit <- nlrq(data = Dat,formula = y ~ SSlogis(x, Asym, mid, scal),tau = 0.5)
summary(fit)

# Help page for function being twiddled
?quantreg::nlrq
}
\seealso{
\code{\link[quantreg]{nlrq}}
}
