% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-stats.R
\name{prcomp}
\alias{prcomp}
\alias{predict.prcomp}
\title{data.frame-first formula-second method for \code{\link[stats]{prcomp}}}
\usage{
prcomp(data, formula = ~., ...)

\method{predict}{prcomp}(object, data, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{a formula with no response variable, referring only to numeric
variables}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[stats]{prcomp}}.
}
\examples{
prcomp(mtcars)

fit <- prcomp(mtcars, ~ .*.)
predict(fit, mtcars[1:5,])
}
\seealso{
\code{\link[stats]{prcomp}}
}
