% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crq.R
\name{predict.twidlr_crq}
\alias{predict.twidlr_crq}
\title{\code{\link[quantreg]{crq}} for twidlr's \code{\link{crq}}}
\usage{
\method{predict}{twidlr_crq}(object, data, ...)
}
\arguments{
\item{object}{Fitted model}

\item{data}{Data frame containing the variables in the model}

\item{...}{Additional arguments to pass to model function}
}
\description{
\code{\link[quantreg]{crq}} for twidlr's \code{\link{crq}}
}
