% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{predict.twidlr_glm}
\alias{predict.twidlr_glm}
\title{\code{\link[stats]{predict.glm}} for twidlr's \code{\link{glm}}}
\usage{
\method{predict}{twidlr_glm}(object, data, ...)
}
\arguments{
\item{object}{Fitted model}

\item{data}{Data frame containing the variables in the model}

\item{...}{Additional arguments to pass to model function}
}
\description{
\code{\link[stats]{predict.glm}} for twidlr's \code{\link{glm}}
}
\examples{
fit <- glm(mtcars, hp ~ .)
predict(fit)
predict(fit, mtcars[1:5, ])
predict(fit, data = mtcars[1:10,])
}
