% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{predict.twidlr_glmnet}
\alias{predict.twidlr_glmnet}
\title{object-first data.frame-second method for \code{\link[glmnet]{predict.glmnet}}}
\usage{
\method{predict}{twidlr_glmnet}(object, data, ...)
}
\description{
object-first data.frame-second method for \code{\link[glmnet]{predict.glmnet}}
}
