% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{predict.twidlr_lm}
\alias{predict.twidlr_lm}
\title{\code{\link[stats]{predict.lm}} for twidlr's \code{\link{lm}}}
\usage{
\method{predict}{twidlr_lm}(object, data, ...)
}
\arguments{
\item{object}{Fitted model}

\item{data}{Data frame containing the variables in the model}

\item{...}{Additional arguments to pass to model function}
}
\description{
\code{\link[stats]{predict.lm}} for twidlr's \code{\link{lm}}
}
\examples{
fit <- lm(mtcars, hp ~ .)
predict(fit)
predict(fit, mtcars[1:5, ])
predict(fit, data = mtcars[1:10,])
}
