% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{predict_checks}
\alias{predict_checks}
\title{Run checks for twidlr predict functions and invisibly return 'data' coerced
to a data frame}
\usage{
predict_checks(data, ...)
}
\arguments{
\item{data}{Value for `data` argument}

\item{...}{Additional named arguments}
}
\description{
Run checks for twidlr predict functions and invisibly return 'data' coerced
to a data frame
}
