% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{rpart}
\alias{rpart}
\title{data.frame-first formula-second method for \code{\link[rpart]{rpart}}}
\usage{
rpart(data, formula, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{A \code{\link[stats]{formula}} with a response but no
interaction terms}

\item{...}{additional arguments to pass to \code{\link[rpart]{rpart}}}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[rpart]{rpart}}.
}
\examples{
fit <- rpart(mtcars, hp ~ .)
summary(fit)

# Help page for function being twiddled
?rpart::rpart
}
\seealso{
\code{\link[rpart]{rpart}}
}
