% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-quantreg.R
\name{rq}
\alias{rq}
\alias{predict.rq}
\title{data.frame-first formula-second method for \code{\link[quantreg]{rq}}}
\usage{
rq(data, formula, ...)

\method{predict}{rq}(object, data, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[quantreg]{rq}}.
}
\examples{
fit <- rq(data = stackloss,formula = stack.loss ~ .,tau = 0.5)
summary(fit)

fit <- rq(data = stackloss,formula = stack.loss ~ .,tau = c(0.25,0.75))
summary(fit)

# Help page for function being twiddled
?quantreg::rq
}
\seealso{
\code{\link[quantreg]{rq}}
}
