% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqss.R
\name{rqss}
\alias{rqss}
\title{data.frame-first formula-second method for \code{\link[quantreg]{rqss}}}
\usage{
rqss(data, formula, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{\code{\link[stats]{formula}} describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[quantreg]{rqss}}.
}
\examples{
n <- 200
x <- sort(rchisq(n,4))
z <- x + rnorm(n)
dat <- data.frame(y = y,x = x,z = z)
rm(x,y,z)
fit  <- rqss(dat,y ~ qss(x, constraint= "N") + z,tau = 0.5)

# Help page for function being twiddled
?quantreg::rqss
}
\seealso{
\code{\link[quantreg]{rqss}}
}
