% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-e1071.R
\name{svm}
\alias{predict.svm}
\alias{svm}
\title{data.frame-first formula-second method for \code{\link[e1071]{svm}}}
\usage{
svm(data, formula = ~., ...)

\method{predict}{svm}(object, data, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[e1071]{svm}}.
}
\seealso{
\code{\link[e1071]{svm}}
}

