% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-stats.R
\name{t.test}
\alias{t.test}
\title{data.frame-first formula-second method for \code{\link[stats]{t.test}}}
\usage{
t.test(data, formula, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{\code{\link[stats]{formula}} describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[stats]{t.test}}.
}
\examples{
t.test(iris[1:100,], Petal.Width ~ Species)

# Help page for function being twiddled
?stats::t.test
}
\seealso{
\code{\link[stats]{t.test}}
}
