% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-stats.R
\name{ttest}
\alias{ttest}
\title{data.frame-first formula-second method for \code{\link[stats]{t.test}}}
\usage{
ttest(data, formula, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[stats]{t.test}}.
}
\examples{
ttest(iris[1:100,], Petal.Width ~ Species)
}
\seealso{
\code{\link[stats]{t.test}}
}

