% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{twiddle}
\alias{twiddle}
\title{Add twidlr_* class and formula attribute to object}
\usage{
twiddle(object, suffix, formula)
}
\arguments{
\item{object}{S3 object}

\item{suffix}{Character string pasted after "twidlr_" and used as new class}

\item{formula}{\code{\link[stats]{formula}} to be added as an attribute "formula"}
}
\description{
Make S3 objects twidlr-compatible to leverage special methods like "predict"
adding a twidlr_* class to it and a formula attribute
}
