% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr.R
\name{twidlr_defaults}
\alias{twidlr_defaults}
\title{Default parameters used by twidlr functions}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{\code{\link[stats]{formula}} describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}
}
\description{
Default parameters used by twidlr functions
}
