% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unpack_data}
\alias{unpack_data}
\title{Convert a data frame and formula to input matrix and output vector}
\usage{
unpack_data(data, f)
}
\arguments{
\item{data}{A data frame}

\item{f}{A formula of the form \code{output ~ feature_1 + feature_2 + ... + feature_i}}
}
\value{
Two-element list containing output values as a vector named \code{y}
  and an input matrix called \code{X}
}
\description{
Convert a data frame and formula of the form "\code{output ~ feature_1 +
feature_2 + ... + feature_i}" to a list containing all input variables as a
matrix and the output variable as a vector. These structures are commonly
used by various modelling functions such as \code{\link[glmnet]{glmnet}}
}
\examples{
unpack_data(mtcars, hp ~ .)
}
