% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unpack_formula}
\alias{unpack_formula}
\title{Convert formula into a character list}
\usage{
unpack_formula(f)
}
\arguments{
\item{f}{A formula}
}
\value{
Two-element list of character vectors corresponding to the lefthand
  and righthand sides of the formula
}
\description{
Convert a formula of the form \code{y ~ x} into a two-element list containing
the lefthand and righthand sides of the formula as character vectors
}
\examples{
unpack_formula(y ~ x)
unpack_formula(y ~ x1 + x2)
unpack_formula(y ~ .)
unpack_formula(~ "x")
}
\seealso{
\code{\link[lazyeval]{f_lhs}}
}
