% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr.R
\name{unsupervised_twidlr_defaults}
\alias{unsupervised_twidlr_defaults}
\title{Default parameters used by twidlr functions that do not specify an outcome
variable (typically an unsupervised model)}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{A \code{\link[stats]{formula}} (or object that can be coerced
to one) with no response variable (optional). If omitted, all variables in
"data" will be included.}

\item{object}{Fitted model}

\item{...}{Additional arguments to pass to model function}
}
\description{
Default parameters used by twidlr functions that do not specify an outcome
variable (typically an unsupervised model)
}

