% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{xgboost}
\alias{xgboost}
\title{data.frame-first formula-second method for xgboost::xgboost}
\usage{
xgboost(data, formula, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model.}

\item{formula}{A formula of the form \code{label ~ feature_1 + feature_2 +
... + feature_i}}

\item{...}{additional arguments to pass to \code{\link[xgboost]{xgboost}}()}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[xgboost]{xgboost}}().
}
\examples{
xgboost(mtcars, hp ~ ., nrounds = 10)
}
\seealso{
\code{\link[xgboost]{xgboost}}
}
