% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{eval_fun}
\alias{eval_fun}
\title{Evaluate a function and list of arguments}
\usage{
eval_fun(.f, ...)
}
\arguments{
\item{.f}{A function}

\item{...}{Arguments to pass to functions}
}
\description{
Evaluate a function and a named list of corresponding arguments
}
\examples{
eval_fun(mean, x = c(1, 2, NA, 3), na.rm = TRUE)
}
