% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-stats.R
\name{factanal}
\alias{factanal}
\alias{predict.factanal}
\title{data.frame-first formula-second method for \code{\link[stats]{factanal}}}
\usage{
factanal(data, formula = ~., ...)

\method{predict}{factanal}(object, data, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{A \code{\link[stats]{formula}} (or object that can be coerced
to one) with no response variable (optional). If omitted, all variables in
"data" will be included.}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[stats]{factanal}}.
}
\examples{
factanal(mtcars, factors = 3)

fit <- factanal(mtcars, ~ hp*am*wt, factors = 3)
predict(fit, mtcars[1:5, ])
}
\seealso{
\code{\link[stats]{factanal}}
}
