% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-stats.R
\name{glm}
\alias{glm}
\alias{predict.glm}
\title{data.frame-first formula-second method for \code{\link[stats]{glm}}}
\usage{
glm(data, formula, ...)

\method{predict}{glm}(object, data, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{\code{\link[stats]{formula}} describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[stats]{glm}}.
}
\examples{
fit <- glm(mtcars, vs ~ hp + wt, family = binomial())
summary(fit)
predict(fit, mtcars[1:5,])

# Help page for function being twiddled
?stats::glm
}
\seealso{
\code{\link[stats]{glm}}
}
