% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-lme4.R
\name{glmer}
\alias{glmer}
\title{data.frame-first formula-second method for \code{\link[lme4]{glmer}}}
\usage{
glmer(data, formula, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[lme4]{glmer}}
}
\examples{
fit <- glmer(lme4::cbpp,
  cbind(incidence, size - incidence) ~ period + (1 | herd),
  family = binomial)
summary(fit)

# Help page for function being twiddled
?lme4::glmer

}
\seealso{
\code{\link[lme4]{glmer}}
}

