% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-glmnet.R
\name{glmnet}
\alias{glmnet}
\alias{predict.glmnet}
\title{data.frame-first formula-second method for \code{\link[glmnet]{glmnet}}}
\usage{
glmnet(data, formula, ...)

\method{predict}{glmnet}(object, data, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{\code{\link[stats]{formula}} describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[glmnet]{glmnet}}().
}
\examples{
glmnet(mtcars, hp ~ .)
}
\seealso{
\code{\link[glmnet]{glmnet}} for source model;
  \code{\link{predict.twidlr_glmnet}} for predictions
}
