% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-stats.R
\name{lm}
\alias{lm}
\alias{predict.lm}
\title{data.frame-first formula-second method for \code{\link[stats]{lm}}}
\usage{
lm(data, formula, ...)

\method{predict}{lm}(object, data, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[stats]{lm}}.
}
\examples{
fit <- lm(mtcars, hp ~ .)
summary(fit)

# Help page for function being twiddled
?stats::lm
}
\seealso{
\code{\link[stats]{lm}}
}

