% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm}
\alias{lm}
\title{data.frame-first formula-second method for \code{\link[stats]{lm}}}
\usage{
lm(data, formula, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{\code{\link[stats]{formula}} describing the model to be fitted}

\item{...}{additional arguments to pass to model function}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[stats]{lm}}.
}
\examples{
fit <- lm(mtcars, hp ~ .)
summary(fit)

# Help page for function being twiddled
?stats::lm
}
\seealso{
\code{\link[stats]{lm}}
}
