% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{predict.twidlr_glmnet}
\alias{predict.twidlr_glmnet}
\title{\code{\link[glmnet]{predict.glmnet}} for twidlr's \code{\link{glmnet}}}
\usage{
\method{predict}{twidlr_glmnet}(object, data, ...)
}
\arguments{
\item{object}{Fitted model}

\item{data}{Data frame containing the variables in the model}

\item{...}{Additional arguments to pass to model function}
}
\description{
\code{\link[glmnet]{predict.glmnet}} for twidlr's \code{\link{glmnet}}
}
\examples{
fit <- glmnet(mtcars, hp ~ .)
coef(fit)
coef(fit, s = 0.01)
predict(fit, data = mtcars)
predict(fit, data = mtcars, s = 0.01)
}
