% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{predict.twidlr_glmnet}
\alias{predict.twidlr_glmnet}
\title{Predict method for \code{\link{glmnet}}}
\usage{
\method{predict}{twidlr_glmnet}(object, data, ...)
}
\arguments{
\item{object}{Fitted model}

\item{data}{Data frame containing the variables in the model}

\item{...}{Additional arguments to pass to model function}
}
\description{
Predict method for \code{\link{glmnet}}
}
