% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{predict.twidlr_glmnet}
\alias{predict.twidlr_glmnet}
\title{twildr compatible predict for \code{\link[glmnet]{glmnet}}}
\usage{
\method{predict}{twidlr_glmnet}(object, data, ...)
}
\description{
twildr compatible predict for \code{\link[glmnet]{glmnet}}
}
