% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{predict.twidlr_lm}
\alias{predict.twidlr_lm}
\title{Predict method for \code{\link{lm}}}
\usage{
\method{predict}{twidlr_lm}(object, data, ...)
}
\arguments{
\item{object}{Fitted model}

\item{data}{Data frame containing the variables in the model}

\item{...}{Additional arguments to pass to model function}
}
\description{
Predict method for \code{\link{lm}}
}
