% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{rpart}
\alias{rpart}
\title{data.frame-first formula-second method for \code{\link[rpart]{rpart}}}
\usage{
rpart(data, formula, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{an object of class \code{"\link{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{...}{additional arguments to pass to \code{\link[rpart]{rpart}}}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[rpart]{rpart}}.
}
\examples{
fit <- rpart(mtcars, hp ~ .)
summary(fit)

# Help page for function being twiddled
?rpart::rpart
}
\seealso{
\code{\link[rpart]{rpart}}
}
