% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{rpart}
\alias{rpart}
\alias{predict.rpart}
\title{data.frame-first formula-second method for \code{\link[rpart]{rpart}}}
\usage{
rpart(data, formula, ...)

\method{predict}{rpart}(object, data, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{\code{\link[stats]{formula}} describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[rpart]{rpart}}.
}
\examples{
rpart(mtcars, hp ~ .)

# Help page for function being twiddled
?rpart::rpart
}
\seealso{
\code{\link[rpart]{rpart}}
}
