% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-stats.R
\name{ttest}
\alias{ttest}
\title{S3 generic method for t.test}
\usage{
ttest(data, formula, ...)
}
\arguments{
\item{data}{Data frame containing the variables in the model}

\item{formula}{\code{\link[stats]{formula}} describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}
}
\description{
S3 generic method for t.test
}
