% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unpack_data}
\alias{unpack_data}
\title{Convert a data frame and formula to input matrix and output vector}
\usage{
unpack_data(data, f)
}
\arguments{
\item{data}{A data frame}

\item{formula}{A formula of the form \code{y ~ x1 + x2 + ... + xi}}
}
\value{
Two-element list containing output values as a vector named \code{y}
  and an input matrix called \code{X}
}
\description{
Convert a data frame and formula of the form "\code{y ~ x1 + x2 + ... + xi}"
to a list containing all input variables as a matrix and the output variable
as a vector. These structures are commonly used by various modelling
functions such as \code{\link[glmnet]{glmnet}}
}
\examples{
unpack_data(mtcars, hp ~ mpg + wt + disp)
unpack_data(mtcars, hp ~ .)
unpack_data(mtcars, ~ .)
}
