% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidlr-xgboost.R
\name{xgboost}
\alias{xgboost}
\alias{predict.xgb.Booster}
\title{data.frame-first formula-second method for \code{\link[xgboost]{xgboost}}}
\usage{
xgboost(data, formula, ...)

\method{predict}{xgb.Booster}(object, data, ...)
}
\arguments{
\item{data}{Data frame (or object that can be coerced to one) containing the
variables in the model}

\item{formula}{\code{\link[stats]{formula}} (or object that can be coerced to
one) describing the model to be fitted}

\item{...}{Additional arguments to pass to model function}

\item{object}{Fitted model}
}
\description{
This function passes a data.frame, formula, and additional arguments to
\code{\link[xgboost]{xgboost}}.
}
\examples{
fit <- xgboost(mtcars, hp ~ ., nrounds = 10, save_period = NULL)
predict(fit, mtcars)
}
\seealso{
\code{\link[xgboost]{xgboost}}
}
